Reorganize DeliveryList class - file loading was relocated to the DataManager class Remove unnecessary imports Rename .delivery.txt Clean up code ready - completed some todos Ready for v2.0 progress
