+++ data/profile.txt
@@ -1,1 +1,1 @@
-Obi-Wan | SJU7606F | BMW X-Wing
+Anakin Skywalker | HIGHGROUND | Jedi Cruiser
