+++ src/main/java/seedu/duke/Duke.java
@@ -1,23 +1,20 @@
 package seedu.duke;
 
-import java.util.Scanner;
-
 public class Duke {
     public Ui ui;
-    public DataManager dataManager;
     public Deliveryman deliveryman;
     /**
      * Main entry-point for the java.duke.Duke application.
      */
     public Duke() {
         ui = new Ui();
-        dataManager = new DataManager();
     }
 
+
     public void run() {
-        deliveryman = dataManager.loadProfile();
+        deliveryman = DataManager.loadProfile();
         ui.showWelcomeScreen();
-        ui.showLoopingMenuUntilExit(deliveryman, dataManager);
+        ui.showLoopingMenuUntilExit(deliveryman);
     }
 
     public static void main(String[] args) {
